bootloader: use system/BCB for bootloader command block (BCB)
Typically a simple structure is stored in a raw partition to facilitate
communication between the bootloader and the recovery application between
power cycles.

Add the ability for this to be stored as a file instead and specify that
file as /system/BCB for our use. This allows the same mechanism for
NAND/ubi and block/ext4 recovery so that they do not need a different
OTA mechanism.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>